%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{as.character.KelkooScan}
\alias{clearScans.KelkooScan}
\alias{generateReport.KelkooScan}
\alias{getAsteriskTags.KelkooScan}
\alias{getFullname.KelkooScan}
\alias{getKelkooDomain}
\alias{getKelkooDomain.KelkooScan}
\alias{getName.KelkooScan}
\alias{getOutputPath.KelkooScan}
\alias{getPostUrl}
\alias{getPostUrl.KelkooFlightScan}
\alias{getPostUrl.KelkooScan}
\alias{getReportPathname.KelkooScan}
\alias{getReportUrl.KelkooScan}
\alias{getRootPath.KelkooScan}
\alias{getScanPathname.KelkooScan}
\alias{getTags.KelkooScan}
\alias{getToday.KelkooScan}
\alias{getVerbose.KelkooScan}
\alias{isOneWay}
\alias{isOneWay.KelkooFlightScan}
\alias{parseHtmlTree.KelkooScan}
\alias{query.KelkooScan}
\alias{readScanTable.KelkooScan}
\alias{setRootPath.KelkooScan}
\alias{setVerbose.KelkooScan}
\alias{as.character.KelkooFlightScan}
\alias{asRow}
\alias{asRow.KelkooItem}
\alias{asStringVector}
\alias{asStringVector.KelkooFlightTicket}
\alias{asStringVector.KelkooItem}
\alias{clearScans}
\alias{generateReport}
\alias{generateReport.KelkooFlightScan}
\alias{getAsteriskTags}
\alias{getCarrierRules}
\alias{getCarrierRules.KelkooFlightScan}
\alias{getDepDates}
\alias{getDepDates.KelkooFlightScan}
\alias{getFrom}
\alias{getFrom.KelkooFlightScan}
\alias{getFromToTag}
\alias{getFromToTag.KelkooFlightScan}
\alias{getFullname}
\alias{getHomeboundItem}
\alias{getHomeboundItem.KelkooFlightTicket}
\alias{getItems}
\alias{getItems.KelkooItem}
\alias{getNbrOfDays}
\alias{getNbrOfDays.KelkooFlightScan}
\alias{getOutboundItem}
\alias{getOutboundItem.KelkooFlightTicket}
\alias{getOutputPath}
\alias{getPrice}
\alias{getPrice.KelkooItem}
\alias{getReportPathname}
\alias{getReportUrl}
\alias{getRetDates}
\alias{getRetDates.KelkooFlightScan}
\alias{getRootPath}
\alias{getRowHeader}
\alias{getRowHeader.KelkooItem}
\alias{getScanPathname}
\alias{getShop}
\alias{getShop.KelkooItem}
\alias{getTags}
\alias{getTo}
\alias{getTo.KelkooFlightScan}
\alias{getToday}
\alias{isRoundTrip}
\alias{isRoundTrip.KelkooFlightTicket}
\alias{KelkooFlightTicket}
\alias{KelkooItem}
\alias{nbrOfItems}
\alias{nbrOfItems.KelkooItem}
\alias{parseHtmlTree}
\alias{print.KelkooItem}
\alias{query}
\alias{queryOneDay}
\alias{queryOneDay.KelkooFlightScan}
\alias{readScanPairs}
\alias{readScanPairs.KelkooFlightScan}
\alias{readScanTable}
\alias{setRootPath}
\alias{setVerbose}
\alias{translateAirfair}
\alias{translateAirfair.KelkooFlightScan}
\alias{translateCarrier}
\alias{translateCarrier.KelkooFlightScan}
\alias{getCurrency}
\alias{getCurrency.KelkooItem}
\alias{getKelkooCountry}
\alias{getKelkooCountry.KelkooScan}
\alias{setKelkooCountry}
\alias{setKelkooCountry.KelkooScan}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
