%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{R.kelkoo-package}
\alias{R.kelkoo-package}
\alias{R.kelkoo}
\docType{package}

\title{Package R.kelkoo}


 \description{
   This package provides methods for querying Kelkoo.com for flights.  One can specify all flight details as on the webpage, but also specify a set of dates for the outbound and the homebound flights to automate the process of searching for optimal flights during a certain period.  The results are given in a tab-delimited file that can be further processed in a spreadsheet application.

   This package should be considered to be in an alpha phase.
   You should expect the API to be changing over time.
 }

 \section{Installation and updates}{

   To install this package, see instructions at
   \url{http://www.braju.com/R/}.
 }

 \section{To get started}{
   To get started, see:
   \enumerate{
     \item \code{\link{KelkooFlightScan}} - See the example.
   }
 }

 \section{How to cite this package}{
 }

 \section{License}{
   The releases of this package is licensed under
   LGPL version 2.1 or newer.
 }

 \references{
 }
\keyword{package}
